+++ web/src/components/ui/index.ts
@@ -387,3 +387,4 @@ export * from './chart-line-wad';
 export * from './chart-line-derivative-osc';
 export * from './chart-line-bop';
 export * from './chart-line-asi';
+export * from './chart-line-gapo';